briefff/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
breadability==0.1.20
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.8.0
docopt==0.6.2
exceptiongroup==1.1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
firebase_admin==6.4.0
flask==2.2.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
jieba3k==0.35.1
joblib==1.3.2
lxml==4.9.3
newspaper3k==0.2.8
nltk==3.8.1
openai==1.2.2
Pillow==10.1.0
pycountry==22.3.5
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sumy==0.11.0
tinysegmenter==0.3
tldextract==5.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
werkzeug==2.2.3
flask == 2.2.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-tasks==2.16.1
metaphor-python==0.1.23
flask_cors==4.0.0
sentence_transformers==2.3.1
pinecone_client==3.0.2

python-dotenv==1.0.1
octoai-sdk==0.8.3
backoff==2.2.1